Added quite a few Nullary classes for the sake of clean construction.
Also created functional Parser interface. Changed name of "Grammar" to
"TerminalSpelling", short extension "Spelling", and introduced
"PhraseStructure" as an analogue for Parser. ("TerminalSpelling" and
"PhraseStructure" could both, ultimately, be built into a single
"Grammar" class in the future; as they would in virtually all cases
refer to the same language.)

Next up is implementing AlgebraicParser, and checking it all out.