Got visitor pattern working (reliably) in Parser, or at least Parser2. Not entirely sure about ordering yet; there could be a bug in SourcePosition. It might be wise to build a JavaFX pane that works the analysis into a nice, clean GUI tree.
I'm sure everyone can imagine me pumping my secret-devil-sign fist in
the air, after the week I've been dealing with. I totally did.

Next up is to verify the ordering and fill in all the blanks. Also,
there're a lot of anonymous classes in AlgebraicPhraseStructure, which
could probably be reduced to a single defined class for the sake of
reducing memory footprint.