Significant cleanup accomplished. Tokens in code relevant to attempted conversion of java.util.regex utilities to work with class sequences have been (to current knowledge) removed. Parser2 has been refactored to Parser, replacing in full the old class called "Parser".
Decision on regex for class sequences: possible to implement; but not
easy and not recommended as of the moment. Regular expressions have
loops available to them that can require significant amounts of time to
process, sometimes going to such a length as to consume all memory. Code
structure should not also have this burden, as it cannot be so easily
remanaged. Visitor pattern is more efficient.